+++ src/Converter/DBAL/NotEqualQueryExprConverter.php
@@ -20,10 +20,9 @@ class NotEqualQueryExprConverter extends AbstractQueryExprConverter
 	public function apply($dbDriver, $alias, $field, $exprValue, $namedParameter = null, &$value = null)
 	{
 		$matches = [];
-		preg_match(self::PATTERN, $exprValue, $matches);
 		
 		if (!preg_match(self::PATTERN, $exprValue, $matches)) {
-			throw new BadQueryExprException(sprintf('The not equal query expression converter does not support the following pattern "%s".', $exprValue));
+			throw new BadQueryExprException(sprintf('The query expression converter "neq" does not support the following pattern "%s".', $exprValue));
 		}
 		
 		$value = $matches[1];